requirements.txt
RPi.GPIO==0.7.1
pygame==2.5.2
adafruit-circuitpython-pn532==2.4.2
rx==3.2.0 
python-dotenv==1.0.0
adafruit-blinka==8.32.0